Update client.java
made the balance updates and requests functional. Made the different game lobbies also create guis for the different games the players can connect to. made functions to get states of the game that are sent by the user like stand and hit and also made the communication method to update status of game (mainly just to get the dealers hand not other players)